fix(across-v2): code review fixes
